feat: Emit rename events with Inotify and Windows
  Since we can follow renames and moves in InotifyBackend and
  WindowsBackend, we can emit rename events rather than a pair of delete
  and create events. This allows applications using @parcel/watcher to
  follow these more easily.

  Using the inode and fileId stored in the snapshot, we can follow
  document renames and moves even when the watcher was not subscribed
  when it happened:

    - If the kind associated with an inode or fileId has changed, we
      assume the document has been replaced and notify a deletion
      followed by a creation.
    - If the inode or fileId cannot be found in the snapshot, we assume
      this is a new document. If another document exists in the snapshot
      at the same path, we notify an update instead.
    - If the inode or fileId cannot be found in the watcher directory,
      we assume the document was deleted and notify it.

+++ test/since.js
@@ -130,10 +130,16 @@ describe('since', () => {
           let res = await watcher.getEventsSince(tmpDir, snapshotPath, {
             backend,
           });
+          if (backend === 'inotify' || backend === 'windows') {
+          assert.deepEqual(res, [
+            event({type: 'rename', oldPath: f1, path: f2, kind, ino, fileId}, {backend}),
+          ]);
+          } else {
           assert.deepEqual(res, [
             event({type: 'delete', path: f1, ino, fileId, kind}, {backend}),
             event({type: 'create', path: f2, ino, fileId, kind}, {backend}),
           ]);
+          }
         });
 
         it('should emit when a file is deleted', async () => {
@@ -182,10 +188,16 @@ describe('since', () => {
             backend,
           });
 
-          assert.deepEqual(res, [
-            event({type: 'delete', path: f1, ino, fileId, kind}, {backend}),
-            event({type: 'create', path: f2, ino, fileId, kind}, {backend}),
-          ]);
+          if (backend === 'inotify' || backend === 'windows') {
+            assert.deepEqual(res, [
+              event({type: 'rename', oldPath: f1, path: f2, ino, fileId, kind}, {backend}),
+            ]);
+          } else {
+            assert.deepEqual(res, [
+              event({type: 'delete', path: f1, ino, fileId, kind}, {backend}),
+              event({type: 'create', path: f2, ino, fileId, kind}, {backend}),
+            ]);
+          }
         });
 
         it('should emit when a directory is deleted', async () => {
@@ -261,10 +273,16 @@ describe('since', () => {
           let res = await watcher.getEventsSince(tmpDir, snapshotPath, {
             backend,
           });
-          assert.deepEqual(res, [
-            event({type: 'delete', path: f2, ino, fileId, kind}, {backend}),
-            event({type: 'create', path: f3, ino, fileId, kind}, {backend}),
-          ]);
+          if (backend === 'inotify' || backend === 'windows') {
+            assert.deepEqual(res, [
+              event({type: 'rename', oldPath: f2, path: f3, ino, fileId, kind}, {backend}),
+            ]);
+          } else {
+            assert.deepEqual(res, [
+              event({type: 'delete', path: f2, ino, fileId, kind}, {backend}),
+              event({type: 'create', path: f3, ino, fileId, kind}, {backend}),
+            ]);
+          }
         });
 
         it('should emit when a sub-file is deleted', async () => {
@@ -320,10 +338,16 @@ describe('since', () => {
           let res = await watcher.getEventsSince(tmpDir, snapshotPath, {
             backend,
           });
-          assert.deepEqual(res, [
-            event({type: 'delete', path: f2, ino, fileId, kind}, {backend}),
-            event({type: 'create', path: f3, ino, fileId, kind}, {backend}),
-          ]);
+          if (backend === 'inotify' || backend === 'windows') {
+            assert.deepEqual(res, [
+              event({type: 'rename', oldPath: f2, path: f3, ino, fileId, kind}, {backend}),
+            ]);
+          } else {
+            assert.deepEqual(res, [
+              event({type: 'delete', path: f2, ino, fileId, kind}, {backend}),
+              event({type: 'create', path: f3, ino, fileId, kind}, {backend}),
+            ]);
+          }
         });
 
         it('should emit when a sub-directory is deleted with files inside', async () => {
@@ -410,10 +434,16 @@ describe('since', () => {
           let res = await watcher.getEventsSince(tmpDir, snapshotPath, {
             backend,
           });
-          assert.deepEqual(res, [
-            event({type: 'delete', path: f2, ino, fileId, kind}, {backend}),
-            event({type: 'create', path: f3, ino, fileId, kind}, {backend}),
-          ]);
+          if (backend === 'inotify' || backend === 'windows') {
+            assert.deepEqual(res, [
+              event({type: 'rename', oldPath: f2, path: f3, ino, fileId, kind}, {backend}),
+            ]);
+          } else {
+            assert.deepEqual(res, [
+              event({type: 'delete', path: f2, ino, fileId, kind}, {backend}),
+              event({type: 'create', path: f3, ino, fileId, kind}, {backend}),
+            ]);
+          }
         });
 
         it('should emit when a symlink is deleted', async () => {
